requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
WeasyPrint==59.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
MarkupSafe==2.1.3
wtforms-sqlalchemy==0.3